test(parser): add comprehensive parser tests and remove old tests
Removed the monolithic parser_tests.rs with 446 lines of tests and replaced it with a new modular suite consisting of legacy_tests.rs, match_tests.rs, mode_tests.rs, and a test module aggregator. Added detailed tests for legacy search syntax, modern match formula syntax, different rule modes, and cross-mode validation. Also refactored the parser functions for collecting unexpected fields for better formatting.
