+++ crates/sempai-yaml/src/tests/parser_tests/legacy_tests.rs
@@ -1,0 +1,105 @@
+//! Tests for legacy search syntax and basic YAML parsing.
+
+use super::*;
+
+#[test]
+fn parse_legacy_search_rule() {
+    let yaml = concat!(
+        "rules:\n",
+        "  - id: demo.legacy\n",
+        "    message: detect foo\n",
+        "    languages: [python]\n",
+        "    severity: WARNING\n",
+        "    pattern: foo($X)\n",
+    );
+
+    check_first_rule(yaml, |rule| {
+        assert_eq!(rule.id(), "demo.legacy");
+        assert_eq!(rule.mode(), &RuleMode::Search);
+        assert_eq!(rule.message(), Some("detect foo"));
+        assert_eq!(rule.languages(), &["python"]);
+        assert_eq!(rule.severity(), Some(&RuleSeverity::Warning));
+        assert!(matches!(
+            rule.principal(),
+            RulePrincipal::Search(SearchQueryPrincipal::Legacy(LegacyFormula::Pattern(pattern)))
+                if pattern == "foo($X)"
+        ));
+    });
+}
+
+#[test]
+fn invalid_yaml_returns_yaml_parse_diagnostic() {
+    let yaml = concat!(
+        "rules:\n",
+        "  - id: bad\n",
+        "    message: oops\n",
+        "    languages: [rust]\n",
+        "    severity: ERROR\n",
+        "    pattern: [",
+    );
+    let (code, _, has_span) = first_err_diagnostic(yaml);
+    assert_eq!(code, DiagnosticCode::ESempaiYamlParse);
+    assert!(has_span);
+}
+
+#[test]
+fn missing_required_field_returns_schema_diagnostic() {
+    let yaml = concat!(
+        "rules:\n",
+        "  - message: detect foo\n",
+        "    languages: [rust]\n",
+        "    severity: ERROR\n",
+        "    pattern: foo($X)\n",
+    );
+    let (code, message, has_span) = first_err_diagnostic(yaml);
+    assert_eq!(code, DiagnosticCode::ESempaiSchemaInvalid);
+    assert!(message.contains("missing required field"));
+    assert!(has_span);
+}
+
+#[rstest]
+#[case::search_rule(concat!(
+    "rules:\n",
+    "  - id: demo.empty.languages\n",
+    "    message: empty languages\n",
+    "    languages: []\n",
+    "    severity: WARNING\n",
+    "    pattern: foo\n",
+))]
+#[case::extract_rule(concat!(
+    "rules:\n",
+    "  - id: demo.empty.languages\n",
+    "    languages: []\n",
+    "    mode: extract\n",
+    "    dest-language: python\n",
+    "    extract: $X\n",
+    "    pattern: foo($X)\n",
+))]
+fn reject_empty_languages(#[case] yaml: &str) {
+    let (code, message, _has_span) = first_err_diagnostic(yaml);
+    assert_eq!(code, DiagnosticCode::ESempaiSchemaInvalid);
+    assert!(message.contains("field `languages` must not be empty"));
+}
+
+#[test]
+fn reject_taint_rule_with_legacy_search_keys() {
+    let yaml = concat!(
+        "rules:\n",
+        "  - id: demo.taint.pattern\n",
+        "    message: taint with pattern\n",
+        "    languages: [python]\n",
+        "    severity: WARNING\n",
+        "    mode: taint\n",
+        "    pattern: foo($X)\n",
+        "    taint:\n",
+        "      sources: []\n",
+        "      sinks: []\n",
+    );
+    let (code, message, _has_span) = first_err_diagnostic(yaml);
+    assert_eq!(code, DiagnosticCode::ESempaiSchemaInvalid);
+    assert!(
+        message
+            .contains("Taint mode rule contains unexpected principal fields: legacy search keys"),
+        "expected error message to contain 'legacy search keys', got '{message}'"
+    );
+}
